cross entropy added

+++ src/mnist.jl
@@ -5,6 +5,6 @@ function readMnist()
     for i in 1:60000
         push!(dataset_x, reshape(c[:, :, i], 784))
     end
-    dataset_y = convert(Array{Float64}, train_y)
+    dataset_y = convert(Array{Int}, train_y)
     return DataSet(dataset_x, dataset_y)
 end
